Revert "Update configuration for `kbd_backlight` LED"
This reverts commit 06c9e4c7ac55ccd6f26c790b6b51143f268dd5a8.

It doesn't feel ideal to have to update sudoers to allow the `kbd_mode`
script to be run without a password...

It's not really a big deal that the keyboard backlight indicator will
break outside of X/i3 because this will rarely matter. The only main
case where this should matter is when unlocking the screen locker.

Also, the `kbd-capslock` trigger updates when capslock is pressed down,
rather than when it is released. This is descyned from the typical
behavior of applying capslock on key release.